requirements.txt
# Usage: pip install -r requirements.txt
# Python 3.8.9

numpy==1.23
psutil==5.9.5
matplotlib>=3.2.2
opencv-python>=4.6.0
Pillow>=7.1.2
scikit-learn==1.3.2
torch==2.0.1
torchvision==0.15.2
tensorflow==2.12.0
# tensorflow-macos==2.12.0
# tensorflow-metal==1.0.1
onnx==1.14.0
onnxsim==0.4.36
onnx2tf==1.15.4
onnxruntime==1.16.3
onnx-graphsurgeon==0.5.2
sng4onnx==1.0.1